requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Django==5.2.9
django-daraja==1.3.0
filetype==1.2.0
fonttools==4.61.0
idna==3.7
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pi_heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.11
six==1.17.0
sqlparse==0.5.4
tqdm==4.67.1
tzdata==2025.2
urllib3==2.6.0
